# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Money Management",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/pengeluaran": {
            "get": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Tampilkan semua data pengeluaran",
                "operationId": "9961de5703beaa6cda54f49963ddbee0",
                "responses": {
                    "200": {
                        "description": "Berhasil menampilkan daftar pengeluaran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pengeluaran"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Simpan data pengeluaran baru",
                "operationId": "79a38b7eb3fa367f3d96b1c4a80f1c65",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tanggal",
                                    "jumlah",
                                    "user_id"
                                ],
                                "properties": {
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-01"
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Belanja bahan"
                                    },
                                    "jumlah": {
                                        "type": "number",
                                        "example": 150000
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data pengeluaran berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pengeluaran"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pengeluaran/{id}": {
            "get": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Tampilkan detail pengeluaran",
                "operationId": "d2cceab258e8bcb9e110018cc2085978",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengeluaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pengeluaran ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pengeluaran"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Perbarui data pengeluaran",
                "operationId": "af6fd5996b0e31a6ba8ceffd844c4425",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengeluaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-02"
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Update catatan"
                                    },
                                    "jumlah": {
                                        "type": "number",
                                        "example": 200000
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pengeluaran"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Hapus data pengeluaran",
                "operationId": "9a619903239b85cdbad7a59cb252d3de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengeluaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data pengeluaran berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
        "/piutang": {
            "get": {
                "tags": [
                    "Piutang"
                ],
                "summary": "Ambil semua data piutang",
                "operationId": "0d1eb5a1f1f8535407be714c945cb3e1",
                "responses": {
                    "200": {
                        "description": "Sukses"
                    }
                }
            },
            "post": {
                "tags": [
                    "Piutang"
                ],
                "summary": "Buat data piutang",
                "operationId": "40c389cfc1be4a61716c63b530ef4d62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Piutang"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/piutang/{id}": {
            "get": {
                "tags": [
                    "Piutang"
                ],
                "summary": "Ambil piutang berdasarkan ID",
                "operationId": "47060607fd1e9466c57fc6c4bbe451ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Piutang"
                ],
                "summary": "Update data piutang",
                "operationId": "4807aa3fbdcec66ed4c3f045213f5777",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Piutang"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Piutang"
                ],
                "summary": "Hapus data piutang",
                "operationId": "4ed24717e8bff61210ad183a5e93b607",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Pengeluaran": {
                "title": "Pengeluaran",
                "description": "Model for Pengeluaran",
                "xml": {
                    "name": "Pengeluaran"
                }
            },
            "Piutang": {
                "title": "Piutang",
                "required": [
                    "tanggal",
                    "jumlah",
                    "debitur",
                    "user_id"
                ],
                "properties": {
                    "id_piutang": {
                        "type": "integer",
                        "example": 1
                    },
                    "tanggal": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-01"
                    },
                    "catatan": {
                        "type": "string",
                        "example": "Catatan piutang"
                    },
                    "jumlah": {
                        "type": "number",
                        "example": 150000
                    },
                    "debitur": {
                        "type": "string",
                        "example": "Nama debitur"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
...
}